+++ NHS Prase/Model/Persistent Model/Machine/_PRWard.h
@@ -39,23 +39,19 @@ extern const struct PRWardFetchedProperties {
 
 
 
-
-@property (nonatomic, strong) NSNumber* id;
-
+@property (nonatomic, strong) NSNumber *id;
 
 
-@property int64_t idValue;
-- (int64_t)idValue;
-- (void)setIdValue:(int64_t)value_;
+@property long long idValue;
+- (long long)idValue;
+- (void)setIdValue:(long long)value_;
 
 //- (BOOL)validateId:(id*)value_ error:(NSError**)error_;
 
 
 
 
-
-@property (nonatomic, strong) NSString* name;
-
+@property (nonatomic, strong) NSString *name;
 
 
 //- (BOOL)validateName:(id*)value_ error:(NSError**)error_;
@@ -64,31 +60,30 @@ extern const struct PRWardFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSSet *concerns;
+@property (nonatomic, strong) NSSet* concerns;
 
 - (NSMutableSet*)concernsSet;
 
 
 
 
-@property (nonatomic, strong) NSSet *goodNotes;
+@property (nonatomic, strong) NSSet* goodNotes;
 
 - (NSMutableSet*)goodNotesSet;
 
 
 
 
-@property (nonatomic, strong) PRHospital *hospital;
+@property (nonatomic, strong) PRHospital* hospital;
 
 //- (BOOL)validateHospital:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSSet *records;
+@property (nonatomic, strong) NSSet* records;
 
 - (NSMutableSet*)recordsSet;
-
 
 
 
@@ -120,8 +115,8 @@ extern const struct PRWardFetchedProperties {
 - (NSNumber*)primitiveId;
 - (void)setPrimitiveId:(NSNumber*)value;
 
-- (int64_t)primitiveIdValue;
-- (void)setPrimitiveIdValue:(int64_t)value_;
+- (long long)primitiveIdValue;
+- (void)setPrimitiveIdValue:(long long)value_;
